final commit

+++ HarmonicNumber/HarmonicNumber.cs
@@ -11,12 +11,11 @@ namespace HarmonicNumber
         public void PrintHarmonicNumber(int number)
         {
             float sumOfHarmonicNumbers = 0;
-            for (int index = 1; index <= number; index++)
+            for (float index = 1; index <= number; index++)
             {
                 sumOfHarmonicNumbers = (sumOfHarmonicNumbers) + (float)(1 / index);
             }
             Console.WriteLine("Sum is : "+ sumOfHarmonicNumbers);
-            //Console.Write("1/" + index + " + ");
         }
     }
 }
